Refs #2872: Locator_t and VendorId_t refactor! RTCP Headers and control messages (not in CMakeLists.txt yet).

+++ include/fastrtps/transport/UDPv4Transport.h
@@ -156,7 +156,7 @@ protected:
                         {return (memcmp(&lhs, &rhs, sizeof(Locator_t)) < 0); } };
 
    //! For both modes, an input channel corresponds to a port.
-   std::map<uint32_t, UDPSocketInfo*> mInputSockets;
+   std::map<uint16_t, UDPSocketInfo*> mInputSockets;
 
    bool IsInterfaceAllowed(const asio::ip::address_v4& ip);
    std::vector<asio::ip::address_v4> mInterfaceWhiteList;
@@ -164,8 +164,8 @@ protected:
    bool OpenAndBindOutputSockets(Locator_t& locator);
    bool OpenAndBindInputSockets(const Locator_t& locator, std::shared_ptr<MessageReceiver> msgReceiver, bool is_multicast);
 
-   eProsimaUDPSocket OpenAndBindUnicastOutputSocket(const asio::ip::address_v4&, uint32_t& port);
-   eProsimaUDPSocket OpenAndBindInputSocket(uint32_t port, bool is_multicast);
+   eProsimaUDPSocket OpenAndBindUnicastOutputSocket(const asio::ip::address_v4&, uint16_t& port);
+   eProsimaUDPSocket OpenAndBindInputSocket(uint16_t port, bool is_multicast);
 
    /** Function to be called from a new thread, which takes cares of performing a blocking receive
    operation on the ReceiveResource
